+++ src/main/resources/templates/user/product/list.html
@@ -1,0 +1,112 @@
+<!DOCTYPE html>
+<html lang="ko" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>상품 목록</title>
+    <style>
+        /* CSS는 변경 사항이 없습니다. */
+        @import url('https://fonts.googleapis.com/css2?family=Nunito:wght@400;600;700&display.swap');
+        :root {
+            --bg-color: #f4f7fa; --card-bg-color: #ffffff; --primary-color: #6a82fb;
+            --primary-hover-color: #5468d4; --secondary-color: #43d9a0; --secondary-hover-color: #35ad80;
+            --delete-color: #ff7b7b; --delete-hover-color: #e66a6a; --text-color: #333;
+            --text-light-color: #777; --border-color: #e9ecef; --shadow-color: rgba(106, 130, 251, 0.15);
+        }
+        * { box-sizing: border-box; margin: 0; padding: 0; }
+        body { font-family: 'Nunito', sans-serif; background-color: var(--bg-color); color: var(--text-color); padding: 2rem; }
+        .container { max-width: 1200px; margin: 0 auto; }
+        .header { display: flex; justify-content: space-between; align-items: center; margin-bottom: 1rem; }
+        .header h1 { font-size: 2rem; font-weight: 700; color: var(--primary-color); }
+        .header-actions a { display: inline-flex; align-items: center; gap: 0.5rem; padding: 0.8rem 1.5rem; border-radius: 50px; border: none; color: white; font-size: 1rem; font-weight: 600; cursor: pointer; transition: background-color 0.2s ease, transform 0.2s ease; text-decoration: none; background-color: var(--secondary-color); }
+        .header-actions a:hover { transform: translateY(-2px); background-color: var(--secondary-hover-color); }
+        .controls-wrapper { display: flex; justify-content: flex-start; align-items: center; margin-bottom: 2rem; flex-wrap: wrap; gap: 1.5rem; }
+        .control-group { display: flex; align-items: center; gap: 0.5rem; }
+        .control-group label { font-weight: 600; color: var(--text-light-color); }
+        .control-group select { padding: 0.6rem; border-radius: 8px; border: 1px solid var(--border-color); font-family: 'Nunito', sans-serif; font-weight: 600; cursor: pointer; background-color: white; appearance: none; padding-right: 2rem; background-image: url('data:image/svg+xml;charset=US-ASCII,%3Csvg%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20width%3D%22292.4%22%2C%22height%22%3A%22292.4%22%3E%3Cpath%20fill%3D%22%236a82fb%22%20d%3D%22M287%2069.4a17.6%2017.6%200%200%200-13-5.4H18.4c-5%200-9.3%201.8-12.9%205.4A17.6%2017.6%200%200%200%200%2082.2c0%205%201.8%209.3%205.4%2012.9l128%20127.9c3.6%203.6%207.8%205.4%2012.8%205.4s9.2-1.8%2012.8-5.4L287%2095c3.5-3.5%205.4-7.8%205.4-12.8%200-5-1.9-9.4-5.4-13z%22%2F%3E%3C%2Fsvg%3E'); background-repeat: no-repeat; background-position: right .7em top 50%; background-size: .65em auto; }
+        .product-list { display: grid; grid-template-columns: repeat(auto-fill, minmax(300px, 1fr)); gap: 1.5rem; }
+        .product-card { background-color: var(--card-bg-color); border-radius: 12px; box-shadow: 0 4px 15px var(--shadow-color); transition: transform 0.3s ease, box-shadow 0.3s ease; display: flex; flex-direction: column; overflow: hidden; }
+        .product-card:hover { transform: translateY(-5px); box-shadow: 0 8px 25px var(--shadow-color); }
+        .product-card > a { text-decoration: none; color: inherit; display: flex; flex-direction: column; flex-grow: 1;} /* 1. 카드 전체를 링크로 만들기 위한 스타일 */
+        .product-image { width: 100%; height: 200px; object-fit: cover; }
+        .product-info { padding: 1rem; flex-grow: 1; display: flex; flex-direction: column; }
+        .product-name { font-size: 1.2rem; font-weight: 700; margin-bottom: 0.5rem; }
+        .product-price { font-size: 1.1rem; font-weight: 600; color: var(--primary-color); margin-top: auto; }
+        .product-actions { background-color: #f8f9fa; padding: 0.75rem 1rem; border-top: 1px solid var(--border-color); }
+        .action-btn-text { width: 100%; padding: 0.75rem; border-radius: 8px; border: none; font-size: 1rem; font-weight: 600; text-align: center; background-color: var(--primary-color); color: white; }
+        .pagination { display: flex; justify-content: center; align-items: center; margin-top: 2rem; gap: 0.5rem; }
+        .pagination a, .pagination span { display: inline-block; padding: 0.5rem 1rem; border-radius: 8px; text-decoration: none; color: var(--text-color); background-color: var(--card-bg-color); border: 1px solid var(--border-color); transition: background-color 0.2s, color 0.2s; }
+        .pagination a:hover { background-color: var(--primary-hover-color); color: white; }
+        .pagination .active { background-color: var(--primary-color); color: white; border-color: var(--primary-color); }
+        .pagination .disabled { color: #ccc; background-color: #f8f8f8; cursor: not-allowed; }
+    </style>
+</head>
+<body>
+
+<div class="container">
+    <header class="header">
+        <h1>상품 목록 🛍️</h1>
+        <div class="header-actions">
+            <a th:href="@{/wishes}">
+                <svg xmlns="http://www.w3.org/2000/svg" width="20" height="20" fill="currentColor" viewBox="0 0 16 16"><path fill-rule="evenodd" d="M8 1.314C12.438-3.248 23.534 4.735 8 15-7.534 4.736 3.562-3.248 8 1.314z"/></svg>
+                <span>위시리스트</span>
+            </a>
+        </div>
+    </header>
+    <div class="controls-wrapper"> </div>
+
+    <main class="product-list">
+        <div class="product-card" th:each="product : ${products.content}">
+            <a th:href="@{/products/{id}(id=${product.id})}">
+                <img th:src="${product.imageUrl}" alt="Product Image" class="product-image">
+                <div class="product-info">
+                    <h2 class="product-name" th:text="${product.name}"></h2>
+                    <p class="product-price" th:text="|${#numbers.formatInteger(product.price, 3, 'COMMA')}원|"></p>
+                </div>
+                <div class="product-actions">
+                    <div class="action-btn-text">옵션 선택하기</div>
+                </div>
+            </a>
+        </div>
+    </main>
+
+    <nav class="pagination" th:if="${products.totalPages > 1}"> </nav>
+</div>
+
+<script th:inline="javascript">
+    /*<![CDATA[*/
+    document.addEventListener('DOMContentLoaded', () => {
+        // --- 3. 불필요한 위시리스트 추가 로직 삭제 ---
+        // 상품 목록 페이지에서는 상세 페이지로 이동하여 옵션을 선택하므로,
+        // 여기서는 위시리스트 추가 fetch 로직이 필요 없습니다.
+
+        // --- 정렬 및 사이즈 변경 JavaScript (기존과 동일) ---
+        const currentPage = /*[[${products.number}]]*/ 0;
+        const currentSize = /*[[${products.size}]]*/ 10;
+        const sortSelector = document.getElementById('sort-selector');
+        if (sortSelector) {
+            sortSelector.addEventListener('change', () => {
+                const currentUrl = new URL(window.location.href);
+                currentUrl.searchParams.set('sort', sortSelector.value);
+                currentUrl.searchParams.set('page', '0');
+                window.location.href = currentUrl.href;
+            });
+        }
+        const sizeSelector = document.getElementById('size-selector');
+        if (sizeSelector) {
+            sizeSelector.addEventListener('change', () => {
+                const currentUrl = new URL(window.location.href);
+                const newSize = sizeSelector.value;
+                const currentOffset = currentPage * currentSize;
+                const newPage = Math.floor(currentOffset / newSize);
+                currentUrl.searchParams.set('size', newSize);
+                currentUrl.searchParams.set('page', newPage.toString());
+                window.location.href = currentUrl.href;
+            });
+        }
+    });
+    /*]]>*/
+</script>
+
+</body>
+</html>
